requirements.txt
azure-functions
azure-storage-blob
pandas
PyMuPDF
requests
msal